find-systype.sh: remove unused "systype" mechanism
The "systype" machanism is merely a system info gatherer, and the info it
collects is (mostly) not used anywhere. The only seemingly important info is
the kernel and kernel version (or the lack thereof), and has been moved to
where it is used (print-ar.sh)

This gets rid of two cross-incompatible buildtime programs.
One tried to find if the build was happening on NeXT, but that's not an
operating system that print-ar.sh cares about. The other tried to do cpuid on
x86, which is useless to the build.

This seems to implement what #11 started.