build system: switch to configure script and improved Makefile
This has dramatic effects on the build system complexity, and in the process
exposing lots of dead weight.

Split header files (.h1 .h2) are merged, and header files that only contained a
single define have been fully replaced with -D. The makefile is no longer
geneating scripts, and their files that hardcoded cc/ld/ar calls are gone.

The build system is now fully compatible with standard variables like CC,
PREFIX, DESTDIR, and such. The compiler is autodetected when unspecified, build
flags are checked for compatibility, and installation directories can now be
specified with configure flags.

The configure script was heavily inspired by XBPS's.

The "compile" script is gone along with the original top-level Makefile, both
replaced with straightforward and more conventional makefiles. The "man" and
"completions" subdirs are now integrated into the build system and use
configure-time variables for the paths.

No sed calls or patches should be required to set any paths or toolchain
settings.

Usage of .gitattributes and "$Id$" strings is now replaced with a VERSION
define in the configure script, which optionally uses git to append the 8-digit
commit hash to the version string. This also makes '-V' more meaningful.