iopause: assume poll() is available
Since poll() is available on all platforms we care about, we don't need to
check if it is.

This gets rid of trypoll.c, the last configure-time check that required to be
run on the host (messing with cross). We also get rid of trysysel.c, since
select() was only used as a fallback when poll() isn't avaialble.